Don't fail if w3af already finished

+++ examples/w3af/pkill_w3af.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
-set -x
+#set -x
 
 # Just in case...
-pkill -9 -f w3af_console
+PID=`ps -eaf | grep w3af_console | grep -v grep | awk '{print $2}'`
+
+if [[ "" !=  "$PID" ]]
+then
+    echo 'Killing w3af_console'
+    pkill -9 -f w3af_console
+else
+    echo 'w3af_console is not running'
+fi
